+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/arguments/UserNameArgument.java
@@ -1,45 +1,46 @@
-package com.nisovin.shopkeepers.commands.lib.arguments;
+package com.nisovin.shopkeepers.commands.arguments;
 
 import java.util.Collections;
 import java.util.Locale;
 import java.util.Objects;
 
-import org.bukkit.entity.Player;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
+import com.nisovin.shopkeepers.api.user.User;
 import com.nisovin.shopkeepers.commands.lib.CommandInput;
 import com.nisovin.shopkeepers.commands.lib.argument.filter.ArgumentFilter;
+import com.nisovin.shopkeepers.commands.lib.arguments.ObjectNameArgument;
 import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
+import com.nisovin.shopkeepers.commands.util.UserArgumentUtils;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.text.Text;
-import com.nisovin.shopkeepers.util.bukkit.EntityUtils;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 import com.nisovin.shopkeepers.util.java.StringUtils;
 
 /**
- * Provides suggestions for names of online players.
+ * Provides suggestions for (potentially offline) {@link User} names.
  * <p>
- * By default this accepts any name regardless of whether it corresponds to an online player.
+ * By default this accepts any name regardless of whether it corresponds to a known user.
  */
-public class PlayerNameArgument extends ObjectNameArgument {
+public class UserNameArgument extends ObjectNameArgument {
 
 	public static final int DEFAULT_MINIMUM_COMPLETION_INPUT = ObjectNameArgument.DEFAULT_MINIMUM_COMPLETION_INPUT;
 
-	// Note: Not providing a default argument filter that only accepts names of online players,
-	// because this can be achieved more efficiently by using PlayerByNameArgument instead.
+	// Note: Not providing a default argument filter that only accepts names of known users,
+	// because this can be achieved more efficiently by using UserByNameArgument instead.
 
-	public PlayerNameArgument(String name) {
+	public UserNameArgument(String name) {
 		this(name, ArgumentFilter.acceptAny());
 	}
 
-	public PlayerNameArgument(String name, ArgumentFilter<? super String> filter) {
+	public UserNameArgument(String name, ArgumentFilter<? super String> filter) {
 		this(name, filter, DEFAULT_MINIMUM_COMPLETION_INPUT);
 	}
 
-	// Joining remaining args doesn't make much sense for player names (and we normalize whitespace
+	// Joining remaining args doesn't make much sense for user names (and we normalize whitespace
 	// in display names).
-	public PlayerNameArgument(
+	public UserNameArgument(
 			String name,
 			ArgumentFilter<? super String> filter,
 			int minimumCompletionInput
@@ -67,18 +68,18 @@ public class PlayerNameArgument extends ObjectNameArgument {
 	 *            the minimum prefix length before completion suggestions are provided
 	 * @param namePrefix
 	 *            the name prefix, may be empty, not <code>null</code>
-	 * @param playerFilter
-	 *            only suggestions for players accepted by this filter are included
+	 * @param userFilter
+	 *            only suggestions for users accepted by this filter are included
 	 * @param includeDisplayNames
 	 *            <code>true</code> to include display name suggestions
-	 * @return the player name completion suggestions
+	 * @return the user name completion suggestions
 	 */
 	public static Iterable<? extends String> getDefaultCompletionSuggestions(
 			CommandInput input,
 			CommandContextView context,
 			int minimumCompletionInput,
 			String namePrefix,
-			ArgumentFilter<? super Player> playerFilter,
+			ArgumentFilter<? super User> userFilter,
 			boolean includeDisplayNames
 	) {
 		// Only provide suggestions if there is a minimum length input:
@@ -90,16 +91,16 @@ public class PlayerNameArgument extends ObjectNameArgument {
 		// color codes).
 		// Normalizes whitespace and converts to lowercase:
 		String normalizedNamePrefix = StringUtils.normalize(namePrefix);
-		Iterable<String> suggestions = EntityUtils.getOnlinePlayersStream()
-				.filter(player -> playerFilter.test(input, context, player))
-				.<@Nullable String>map(player -> {
-					// Note: Not suggesting both the name and display name for the same player.
-					// Assumption: Player names don't contain whitespace or color codes
-					String name = Unsafe.assertNonNull(player.getName());
+		Iterable<String> suggestions = UserArgumentUtils.getKnownUsers()
+				.filter(user -> userFilter.test(input, context, user))
+				.<@Nullable String>map(user -> {
+					// Note: Not suggesting both the name and display name for the same user.
+					// Assumption: User names don't contain whitespace or color codes
+					String name = Unsafe.assertNonNull(user.getName());
 					if (StringUtils.normalize(name).startsWith(normalizedNamePrefix)) {
 						return name;
 					} else if (includeDisplayNames) {
-						String displayName = TextUtils.stripColor(player.getDisplayName());
+						String displayName = TextUtils.stripColor(user.getDisplayName());
 						String normalizedWithCase = StringUtils.normalizeKeepCase(displayName);
 						String normalized = normalizedWithCase.toLowerCase(Locale.ROOT);
 						if (normalized.startsWith(normalizedNamePrefix)) {